Notification deletion on confirm click, error check for requesting facil

+++ 395project/395project/dash/RequestFacilitator.aspx.cs
@@ -30,29 +30,49 @@ namespace _395project.Pages
             SqlConnection conn = new SqlConnection(ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
             conn.Open();
             string insert = "insert into RequestFacilitator(Email, FacilitatorFirstName, FacilitatorLastName) values (@CurrentUser, @FacilitatorFirst, @FacilitatorLast)";
-            SqlCommand cmd = new SqlCommand(insert, conn);
-            cmd.Parameters.AddWithValue("@CurrentUser", User.Identity.GetUserId());
-            cmd.Parameters.AddWithValue("@FacilitatorFirst", FacilitatorFirst.Text);
-            cmd.Parameters.AddWithValue("@FacilitatorLast", FacilitatorLast.Text);
+            string check = "SELECT COUNT(*) FROM Facilitators WHERE (Id = @CurrentUser and FirstName = @FirstName and LastName = @LastName)";
+            SqlCommand checkExists = new SqlCommand(check, conn);
+            checkExists.Parameters.AddWithValue("@CurrentUser", User.Identity.GetUserId());
+            checkExists.Parameters.AddWithValue("@FirstName", FacilitatorFirst.Text);
+            checkExists.Parameters.AddWithValue("@LastName", FacilitatorLast.Text);
+            int facilitatorExists = (int)checkExists.ExecuteScalar();
 
-            cmd.ExecuteNonQuery();
+            if(facilitatorExists > 0)
+            {
+                ErrorMessages.Visible = true;
+                ErrorMessages.ForeColor = System.Drawing.Color.Red;
+                ErrorMessages.Text = "Facilitator already associated with your account!";
+                conn.Close();
+            }
+            else
+            {
+                SqlCommand cmd = new SqlCommand(insert, conn);
+                cmd.Parameters.AddWithValue("@CurrentUser", User.Identity.GetUserId());
+                cmd.Parameters.AddWithValue("@FacilitatorFirst", FacilitatorFirst.Text);
+                cmd.Parameters.AddWithValue("@FacilitatorLast", FacilitatorLast.Text);
 
-            string remove = "delete from RequestFacilitator where Email = '' or FacilitatorFirstName = '' or FacilitatorLastName = ''";
-            SqlCommand rm = new SqlCommand(remove, conn);
-            rm.ExecuteNonQuery();
-            conn.Close();
+                cmd.ExecuteNonQuery();
+
+                string remove = "delete from RequestFacilitator where Email = '' or FacilitatorFirstName = '' or FacilitatorLastName = ''";
+                SqlCommand rm = new SqlCommand(remove, conn);
+                rm.ExecuteNonQuery();
+                conn.Close();
+
+                ErrorMessages.Visible = true;
+                ErrorMessages.ForeColor = System.Drawing.Color.Green;
+                ErrorMessages.Text = "Facilitator Request Sent!";
+            }
 
             FacilitatorFirst.Text = string.Empty;
             FacilitatorLast.Text = string.Empty;
-            ErrorMessage.Visible = true;
-            ErrorMessage.Text = "Request Facilitator Sent";
+
         }
 
         protected void Clear_Click(object sender, EventArgs e)
         {
             FacilitatorFirst.Text = string.Empty;
             FacilitatorLast.Text = string.Empty;
-            ErrorMessage.Visible = false;
+            ErrorMessages.Visible = false;
         }
     }
 }
